Abstracted out the GetDateRange function to handle getting the days between the months
